Femora/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
cryptography==41.0.7
google-cloud-storage==2.10.0
numpy==1.24.3
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
opencv-python==4.8.1.78

